Setup DOM environment synchronously in setup.js
The 'document' and 'window' globals must be set up
before React is required for the first time.

This happened to work previously but was liable to break
since the async jsdom.env() API was being used.

This resolves the issue by using the synchronous jsdom.jsdom()
API instead in setup.js which is required before any tests are run.

Thanks to @p14n for the report.