Update to React v0.14 and Babel 6
This brings the example up to date with the current versions
of these libraries.

It also revealed an inherent flaw with the approach taken
to stubbing modules - namely that it relied on the the way
that Babel 5 _happened_ to translate ES2015 module syntax to ES5.

Babel 6 uses a different transpilation of default exports
which reveals the problem. The correct solution is to use
a different approach to stubbing modules, which I'll look into
as a follow-up. This commit uses a horrific hack so that the tests
do at least pass.

TweetList.js is unfortunately reformatted in this commit
because the original code uses tabs rather than the conventional
2-space formatting.